09-11-2020

+++ Controllers/TitulosController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
@@ -7,21 +6,24 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using app02.Data;
 using app02.Models.Entidades;
+using app02.Services;
 
 namespace app02.Controllers
 {
     public class TitulosController : Controller
     {
         private readonly app02Context _context;
+        private readonly ClienteService _clienteService;
 
-        public TitulosController(app02Context context)
+        public TitulosController(app02Context context, ClienteService clienteService)
         {
             _context = context;
+            _clienteService = clienteService;
         }
 
         // GET: Titulos
         public async Task<IActionResult> Index()
-        
+
         {
             var app02Context = _context.Titulos.Include(t => t.Cliente);
 
@@ -36,12 +38,11 @@ namespace app02.Controllers
                 return NotFound();
             }
 
-            var titulo = await _context.Titulos
-                .Include(t => t.Cliente)
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (titulo == null)
+            var titulo = await _context.Titulos.Include(c => c.Cliente).FirstOrDefaultAsync(x => x.Id == id);
+
+            if (titulo == null || titulo.Id != id)
             {
-                return NotFound();
+                return NotFound("Titulo Não Encontrado");
             }
 
             return View(titulo);
@@ -59,18 +60,17 @@ namespace app02.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Documento,Valor,Emissao,Vencimento,ClienteId,Status")] Titulo titulo)
+        public async Task<IActionResult> Create(Titulo titulo)
         {
-            if (!ModelState.IsValid)
-            {
-                return NotFound();
-            }
 
-            _context.Add(titulo);
+            titulo.Status = 1;
+            _context.Titulos.Add(titulo);
+
+            await _clienteService.AddTitulosAsync(titulo.ClienteId, titulo);
+
             await _context.SaveChangesAsync();
 
-            ViewData["ClienteId"] = new SelectList(_context.Cliente, "Id", "Id", titulo.ClienteId);
-            return View(titulo);
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Titulos/Edit/5
@@ -78,52 +78,44 @@ namespace app02.Controllers
         {
             if (id == null)
             {
-                return NotFound();
+                return NotFound("Titulo Não Encontrado");
             }
 
             var titulo = await _context.Titulos.FindAsync(id);
-            if (titulo == null)
-            {
-                return NotFound();
-            }
-            ViewData["ClienteId"] = new SelectList(_context.Cliente, "Id", "Id", titulo.ClienteId);
-            return View(titulo);
-        }
 
-        // POST: Titulos/Edit/5
+
+            if (titulo == null || titulo.Id != id)
+            {
+                return NotFound("Titulo Não Encontrado");
+            }
+
+            return View(titulo);
+
+        }
+        // POST: Titulos/Edit
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Documento,Valor,Emissao,Vencimento,ClienteId,Status")] Titulo titulo)
+        public async Task<IActionResult> Edit([Bind("Id,Documento,ClienteId,Emissao,Vencimento,Valor,Status")] Titulo titulo)
         {
-            if (id != titulo.Id)
+
+            if (!ModelState.IsValid)
             {
-                return NotFound();
+                return NotFound("Model State Invalido");
             }
 
-            if (ModelState.IsValid)
+            try
             {
-                try
-                {
-                    _context.Update(titulo);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!TituloExists(titulo.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                _context.Titulos.Update(titulo);
+                await _context.SaveChangesAsync();
             }
-            ViewData["ClienteId"] = new SelectList(_context.Cliente, "Id", "Id", titulo.ClienteId);
-            return View(titulo);
+            catch (DbUpdateConcurrencyException e)
+            {
+                return NotFound(e.Message);
+            }
+
+            return RedirectToAction(nameof(Index));
         }
 
         // GET: Titulos/Delete/5
@@ -137,9 +129,9 @@ namespace app02.Controllers
             var titulo = await _context.Titulos
                 .Include(t => t.Cliente)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (titulo == null)
+            if (titulo == null || titulo.Id != id)
             {
-                return NotFound();
+                return NotFound("Titulo Não Encontrado");
             }
 
             return View(titulo);
@@ -151,18 +143,28 @@ namespace app02.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var titulo = await _context.Titulos.FindAsync(id);
+
+            if (titulo == null || titulo.Id != id)
+            {
+                return NotFound("Titulo Não Encontrado");
+            }
+
             _context.Titulos.Remove(titulo);
+
+            await _clienteService.RemTitulosAsync(titulo.ClienteId, titulo);
+
             await _context.SaveChangesAsync();
+
             return RedirectToAction(nameof(Index));
         }
 
         private bool TituloExists(int id)
         {
-            return _context.Titulos.Any(e => e.Id == id);
+            return _context.Titulos.Any(x => x.Id == id);
         }
 
         // Busca com filtro de Datas de Vencimento
-        public async Task<IActionResult> FindbyDate(DateTime? dataini,DateTime? datafim, Cliente cliente)
+        public async Task<IActionResult> FindbyDate(DateTime? dataini, DateTime? datafim, Cliente cliente)
 
         {
             if (!dataini.HasValue)
@@ -172,7 +174,7 @@ namespace app02.Controllers
 
             if (!datafim.HasValue)
             {
-                dataini = new DateTime(DateTime.Now.Year, 12, 31);
+                datafim = new DateTime(DateTime.Now.Year, 12, 31);
             }
 
 
@@ -186,5 +188,79 @@ namespace app02.Controllers
             return View(await selecao.ToListAsync());
         }
 
+
+        // GET: Titulos/Pagamento
+        public async Task<IActionResult> Payment(int? id)
+        {
+            if (id == null)
+            {
+                return NotFound("Numero do Título Nulo no Metodo Get");
+            }
+
+            var titulo = await _context.Titulos
+              .Include(t => t.Cliente)
+              .FirstOrDefaultAsync(m => m.Id == id);
+
+            if (titulo == null)
+            {
+                return NotFound("Titulo Não Encontrado no Metodo Get");
+            }
+            if (id != titulo.Id)
+            {
+                return NotFound("Titulo Encontrado Diferente do Procurado no Metodo GET");
+            }
+
+            return View(titulo);
+        }
+
+        // POST: Titulos/Pagamento
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost, ActionName("Payment")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Payment(int id, [Bind("Id,Documento,Valor,Emissao,Vencimento,ClienteId,Pagamento,Status")]
+                                                 Titulo titulo)
+        {
+            if (id != titulo.Id)
+            {
+                return NotFound("Titulo Não Encontrado");
+            }
+
+            if (!ModelState.IsValid)
+            {
+                return NotFound("Model State Invalido");
+            }
+
+            try
+            {
+                var tituloatualizado =  CalculaPagamento(titulo);
+                _context.Titulos.Update(tituloatualizado);
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException e)
+            {
+                if (!TituloExists(titulo.Id))
+                {
+                    return NotFound("Título não Existe");
+                }
+                else
+                {
+                    return NotFound(e.Message);
+                }
+            }
+            return RedirectToAction(nameof(Index));
+        }
+        public Titulo CalculaPagamento(Titulo titulo)
+        {
+            if (titulo.Pagamento > titulo.Vencimento)
+            {
+                var atraso = titulo.Pagamento.Subtract(titulo.Vencimento).TotalDays;
+                titulo.Juros = titulo.Valor * 0.01 / 30 * atraso;
+                titulo.Multa = titulo.Valor * 0.02;
+            }
+            titulo.Totalpago = titulo.Valor + titulo.Juros + titulo.Multa;
+            titulo.Status = 2;
+            return titulo;
+        }
     }
 }
